Simplify and optimize Schrage's algorithm by avoiding unnecessary sorting steps.
This changes the behavior in that ties between algorithms with identical cooldown times are broken
by scheduling the job with the longer processing time first.
This new behavior is in accordance to e.g. the "Handbook of Scheduling".
Previously, the job with the shorter processing time was scheduled first; this was likely unintended.